Move implementation into cpp file
